Assiging new values for F_ASSOCI_PID and F_DASSOC_PID to avoid a conflict with the mixer audio driver. It seamd to be that driver is doing weird things when getting ioctl from the kernel.

+++ usr/src/uts/common/sys/fcntl.h
@@ -181,8 +181,8 @@ extern "C" {
 #endif	/* _STRICT_SYMBOLS */
 #endif /* _LP64 || _FILE_OFFSET_BITS == 32 */
 
-#define	F_ASSOCI_PID	(INT_MAX)
-#define	F_DASSOC_PID	(INT_MAX - 1)
+#define	F_ASSOCI_PID	(USHRT_MAX)
+#define	F_DASSOC_PID	(USHRT_MAX - 1)
 
 #if 	defined(_LARGEFILE64_SOURCE)
